create Controller.py litterally just copy pasted the code that fiona wrote for the stragegy pattern in display to the new controller file and it worked perfectly
