+++ YFinanceAdaptor.py
